refactor endpoint rule branch trace
